Добавлен класс Game. Обновлены поля в классах Point, Player.

+++ src/main/java/Point.java
@@ -1,8 +1,9 @@
 /**
- *  Используется для описания точки на плоскости.
+ * Используется для описания точки на плоскости.
  */
 public class Point {
 
-    private int x;
-    private int y;
+    private short x; // value in the range 1 to 10000
+    private short y; // value in the range 1 to 10000
+
 }
